refactor: 특정 주차의 면접 가능 지원자 계산하는 로직 공통화
