+++ src/pages/Interview/components/AvailableTimesMode/index.tsx
@@ -3,6 +3,7 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 import { AvailableTimesHeader } from '@/pages/Interview/components/AvailableTimesMode/AvailableTimesHeader';
 import { InterviewPageLayout } from '@/pages/Interview/components/InterviewPageLayout';
 import { useInterviewPartSelectionContext } from '@/pages/Interview/context';
+import { useAvailableApplicantsInWeek } from '@/pages/Interview/hooks/useAvailableApplicantsInWeek';
 import { useWeekIndicator } from '@/pages/Interview/hooks/useWeekIndicator';
 import { applicantOptions } from '@/query/applicant/options';
 
@@ -11,15 +12,20 @@ export const AvailableTimesMode = () => {
   const { year, month, week, handlePrevWeek, handleNextWeek } = useWeekIndicator();
 
   const { data: applicants } = useSuspenseQuery(applicantOptions({ partId }));
+  const availableApplicants = useAvailableApplicantsInWeek({
+    year,
+    month,
+    week,
+    applicants,
+  });
 
   return (
     <InterviewPageLayout
       slots={{
         header: (
           <AvailableTimesHeader
-            applicants={applicants}
+            availableApplicants={availableApplicants}
             indicator={{
-              year,
               month,
               week,
               onNextWeek: handleNextWeek,
